Simplify data preprocess
